punctuation validator modified, added recordtype parameter and fixed validating
